Implements npm web login support (#6981)
## What's the problem this PR addresses?

Npm plans to fully remove the `/-/user/org.couchdb.user:<name>`
endpoint. Since we use this endpoint to login, we need to implement
support for the newer `/-/v1/login` endpoint.

## How did you fix it?

A new option, `--web-login`, enables the web login flow (and fallbacks
to the password flow if the web flow isn't supported). This flag is
automatically enabled for the npm registry. This is technically a
breaking change, but since it's the registry that plans to make that
change I think it's ok.

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.